Improve doxygen groups

+++ examples/midi-notes.cpp
@@ -13,7 +13,7 @@ auto main() -> int
     for (auto original : notes) {
         auto const hertz = mc::dsp::toHertz<float>(original);
         auto const note  = mc::dsp::toMidiNote(hertz);
-        std::printf("note: %d\tfreq: %f\n", (int)note, hertz.count());
+        std::printf("note: %d\tfreq: %f\n", static_cast<uint8_t>(note), hertz.count());
     }
     return 0;
 }
